Refs #2287. Fixed Windows tests linking error.

+++ test/UtilsTests.cpp
@@ -36,7 +36,7 @@ class TokenBucketTests : public ::testing::Test
 TEST_F(TokenBucketTests, RateNoBurst)
 {
     // Rate and capacity should be automatically adjusted to be equal.
-    const int rate = 70;
+    const unsigned int rate = 70;
     TokenBucket bucket{rate,0};
     ASSERT_TRUE(bucket.get_tokens(rate));
     ASSERT_FALSE(bucket.get_tokens(10));
@@ -52,7 +52,7 @@ TEST_F(TokenBucketTests, RateNoBurst)
 
 TEST_F(TokenBucketTests, NoRateNoBurst)
 {
-    const int min_rate = 64000;
+    const unsigned int min_rate = 64000;
     TokenBucket bucket{0};
     ASSERT_TRUE(bucket.get_tokens(min_rate));
     ASSERT_FALSE(bucket.get_tokens(10));
@@ -68,8 +68,8 @@ TEST_F(TokenBucketTests, NoRateNoBurst)
 
 TEST_F(TokenBucketTests, AdjustedBurst)
 {
-    const int min_rate = 64000;
-    const int rate     = min_rate * 0.5;
+    const unsigned int min_rate = 64000;
+    const unsigned int rate     = (unsigned int) (min_rate * 0.5);
     TokenBucket bucket{rate, 10};
     ASSERT_TRUE(bucket.get_tokens(rate));
     ASSERT_FALSE(bucket.get_tokens(10));
@@ -77,7 +77,7 @@ TEST_F(TokenBucketTests, AdjustedBurst)
     ASSERT_TRUE(bucket.get_tokens(rate * 2));
     ASSERT_FALSE(bucket.get_tokens(1));
     std::this_thread::sleep_for(std::chrono::seconds(1));
-    ASSERT_TRUE(bucket.get_tokens(rate * 0.5));
+    ASSERT_TRUE(bucket.get_tokens((unsigned int) (rate * 0.5)));
     ASSERT_FALSE(bucket.get_tokens(rate));
 }
 
@@ -99,8 +99,8 @@ TEST_F(TokenBucketTests, LimitToUDPBucket)
 TEST_F(TokenBucketTests, RateMeassure)
 {
     const size_t test_package = 300;
-    const int bucket_size     = 100;
-    const size_t tokens       = 50;
+    const unsigned int bucket_size = 100;
+    const size_t tokens = 50;
     TokenBucket bucket{bucket_size,0};
 
     std::cout << "Testing package size: " << test_package << " Rate: " << bucket_size << " Token: " << tokens
